requirements.txt
langchain==0.0.267
openai==0.27.8
chromadb==0.4.6
tiktoken==0.4.0
sentence-transformers==2.2.2
PyPDF2==3.0.1
pyyaml==6.0.1
fastapi==0.100.0
uvicorn==0.23.1
